requirements.txt
alembic==1.2.1
astroid==2.2.5
blinker==1.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
pkg-resources==0.0.0
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.14.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.8
typed-ast==1.4.0
visitor==0.1.3
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
